treewide: Change DOUBLEs used for aggregation into DECIMAL(38, 2).
Sums of DOUBLE are imprecise and cannot be implemented linearly without
extending precision.  Although that is possible, e.g. see
https://github.com/feldera/feldera/issues/4360, it is even easier to use
DECIMAL instead, in many cases.  This commit changes all of the instances
of DOUBLE that are potentially used for aggregation into DECIMAL.

Some DOUBLE values cannot sensibly be used for arithmetical aggregation.
This commit does not change them.  (For example, it doesn't make sense
to take the sum, average, or standard deviation over latitude and
longitude values.)

+++ demo/packaged/sql/04-feature-engineering.sql
@@ -56,7 +56,7 @@ CREATE TABLE customer (
 CREATE TABLE transaction (
     -- Lateness annotation: transactions cannot arrive more than 1 day out of order.
     ts TIMESTAMP LATENESS INTERVAL 1 DAYS,
-    amt DOUBLE,
+    amt DECIMAL(38, 2),
     customer_id BIGINT NOT NULL,
     state VARCHAR
 ) WITH (
